1번회원을관리자로, 게시물 수정/삭제시 권한체크

+++ src/main/java/com/sbs/untact/controller/UsrArticleController.java
@@ -77,7 +77,13 @@ public class UsrArticleController {
 
 	@RequestMapping("usr/article/doDelete")
 	@ResponseBody
-	public ResultData doDelete(Integer id) {
+	public ResultData doDelete(Integer id, HttpSession session) {
+		int loginedMemberId = Util.getAsInt(session.getAttribute("loginedMemberId"), 0);
+
+		if (loginedMemberId == 0) {
+			return new ResultData("F-2", "로그인 후 이용해주세요.");
+		}
+
 		if (id == null) {
 			return new ResultData("F-1", "id를 입력해주세요");
 		}
@@ -87,12 +93,24 @@ public class UsrArticleController {
 			return new ResultData("F-1", "해당 게시물은 존재하지않습니다.");
 		}
 
+		ResultData actorCanDeleteRd = articleService.getActorCanDeleteRd(article, loginedMemberId);
+
+		if (actorCanDeleteRd.isFail()) {
+			return actorCanDeleteRd;
+		}
+
 		return articleService.deleteArticle(id);
 	}
 
 	@RequestMapping("usr/article/doModify")
 	@ResponseBody
-	public ResultData doModify(Integer id, String title, String body) {
+	public ResultData doModify(Integer id, String title, String body, HttpSession session) {
+		int loginedMemberId = Util.getAsInt(session.getAttribute("loginedMemberId"), 0);
+
+		if (loginedMemberId == 0) {
+			return new ResultData("F-2", "로그인 후 이용해주세요.");
+		}
+
 		if (id == null) {
 			return new ResultData("F-1", "id를 입력해주세요");
 		}
@@ -106,6 +124,12 @@ public class UsrArticleController {
 
 		if (article == null) {
 			return new ResultData("F-1", "게시물이 존재하지않습니다.");
+		}
+
+		ResultData actorCanModifyRd = articleService.getActorCanModifyRd(article, loginedMemberId);
+
+		if (actorCanModifyRd.isFail()) {
+			return actorCanModifyRd;
 		}
 
 		return articleService.modifyArticle(id, title, body);
